refactor: click interactivity working again!
